move GHttpRequestX to original place
