+++ src/IIIFPresentation/API/Features/Manifest/ManifestController.cs
@@ -4,7 +4,6 @@ using API.Auth;
 using API.Features.Manifest.Requests;
 using API.Features.Manifest.Validators;
 using API.Features.Storage.Helpers;
-using API.Helpers;
 using API.Infrastructure;
 using API.Infrastructure.Filters;
 using API.Infrastructure.Helpers;
