+++ src/IIIFPresentation/API/Features/Storage/CollectionController.cs
@@ -5,10 +5,10 @@ using API.Features.Storage.Helpers;
 using API.Features.Storage.Models;
 using API.Features.Storage.Requests;
 using API.Features.Storage.Validators;
-using API.Helpers;
 using API.Infrastructure;
 using API.Infrastructure.Filters;
 using API.Infrastructure.Helpers;
+using API.Infrastructure.Requests;
 using API.Settings;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
