+++ src/IIIFPresentation/API/Features/Storage/StorageController.cs
@@ -4,10 +4,10 @@ using API.Auth;
 using API.Converters;
 using API.Features.Manifest.Requests;
 using API.Features.Storage.Requests;
-using API.Helpers;
 using API.Infrastructure;
 using API.Infrastructure.Filters;
 using API.Infrastructure.Helpers;
+using API.Infrastructure.Requests;
 using API.Settings;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
