+++ src/IIIFPresentation/API/Infrastructure/ControllerBaseX.cs
@@ -1,7 +1,6 @@
 using System.Net;
 using System.Runtime.InteropServices.JavaScript;
 using API.Features.Storage.Helpers;
-using API.Helpers;
 using API.Infrastructure.Requests;
 using Core;
 using Core.Helpers;
