+++ src/IIIFPresentation/API/Infrastructure/PresentationController.cs
@@ -1,6 +1,5 @@
 ﻿using System.Net;
 using API.Exceptions;
-using API.Helpers;
 using API.Infrastructure.Requests;
 using API.Settings;
 using Core;
